refactor: Enhance user types structure
