+++ apps/backend/src/types/customer/index.ts
@@ -1,1 +1,2 @@
 export * from "./user";
+export * from "./auth";
